WinTerm: use names instead of chars in settings
It appears that using "period", "comma" and "minus" instead of ".", ","
and "-" makes the shortcuts work the first time Windows Terminal is
open, which is not the case with the characters. Unfortunately, no name
exist for "\" and "|", which means splitting vertically does not work
the first time a WinTerm is opened.